pyproject.toml
[tool.poetry]
name = "testebbm"
version = "0.1.0"
description = "Teste para a BBM"
authors = ["Cristiano Cardoso"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
drf-spectacular = "^0.22.1"
django-redis = "^5.2.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
